# backend/OpenAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "health"
        ],
        "operationId": "home",
        "responses": {
          "200": {
            "description": "Home endpoint"
          }
        }
      }
    },
    "/checks/": {
      "get": {
        "tags": [
          "checks"
        ],
        "summary": "List my checks",
        "description": "Lists all checks the authenticated user has access to",
        "operationId": "listMyChecks",
        "responses": {
          "200": {
            "description": "List of checks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheckWithAccess"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "checks"
        ],
        "summary": "Create a new check",
        "description": "Creates a new check across multiple regions. The creator automatically gets full access (can_edit and can_see).",
        "operationId": "createCheck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Check"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Check created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Check"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/checks/{check_id}": {
      "get": {
        "tags": [
          "checks"
        ],
        "summary": "Get check by ID",
        "description": "Retrieves a check by its ID. User must have access to view the check.",
        "operationId": "getCheck",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Check found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckWithAccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Forbidden - no access to this check"
          },
          "404": {
            "description": "Check not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "checks"
        ],
        "summary": "Delete check",
        "description": "Deletes a check from all regions. User must have edit access to the check.",
        "operationId": "deleteCheck",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Check deleted successfully"
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Forbidden - no edit access to check"
          },
          "404": {
            "description": "Check not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "checks"
        ],
        "summary": "Update check",
        "description": "Updates a check. User must have edit access to the check.",
        "operationId": "updateCheck",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Check"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Check updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Check"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Forbidden - no edit access to check"
          },
          "404": {
            "description": "Check not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/checks/{check_id}/metrics": {
      "get": {
        "tags": [
          "checks"
        ],
        "summary": "Get check metrics",
        "description": "Get aggregated uptime and performance metrics for a check over a time range",
        "operationId": "getCheckMetrics",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "description": "Check ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start timestamp (ISO 8601)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End timestamp (ISO 8601, exclusive)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "regions",
            "in": "query",
            "description": "Comma-separated list of regions to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metrics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "403": {
            "description": "Forbidden - no access to check"
          },
          "404": {
            "description": "Check not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Health check"
          }
        }
      }
    },
    "/internal": {
      "post": {
        "tags": [
          "internal"
        ],
        "operationId": "internal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Internal endpoint success"
          },
          "401": {
            "description": "Unauthorized - invalid or missing password"
          }
        },
        "security": [
          {
            "internal_bearer": []
          }
        ]
      }
    },
    "/users/info/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user by ID",
        "description": "Retrieves a user's public information by their unique identifier",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Login user",
        "description": "Authenticates a user and establishes a session",
        "operationId": "loginUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Logout user",
        "description": "Logs out the current user and invalidates their session",
        "operationId": "logoutUser",
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get current authenticated user",
        "description": "Retrieves the currently authenticated user's public information. Returns null if not authenticated.",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Success - returns user info or null if not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/PublicUser"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/users/new": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a new user",
        "description": "Creates a new user account and establishes a session",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Check": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CheckData"
          },
          {
            "type": "object",
            "required": [
              "check_id",
              "regions"
            ],
            "properties": {
              "check_id": {
                "type": "string",
                "format": "uuid"
              },
              "regions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          }
        ]
      },
      "CheckAccess": {
        "type": "object",
        "required": [
          "can_edit",
          "can_see"
        ],
        "properties": {
          "can_edit": {
            "type": "boolean"
          },
          "can_see": {
            "type": "boolean"
          }
        }
      },
      "CheckData": {
        "type": "object",
        "required": [
          "check_name",
          "url",
          "http_method",
          "check_frequency_seconds",
          "timeout_seconds",
          "expected_status_code",
          "request_headers",
          "is_enabled",
          "created_at"
        ],
        "properties": {
          "check_frequency_seconds": {
            "type": "integer",
            "format": "int32"
          },
          "check_name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expected_status_code": {
            "type": "integer",
            "format": "int32"
          },
          "http_method": {
            "$ref": "#/components/schemas/Method"
          },
          "is_enabled": {
            "type": "boolean"
          },
          "request_body": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "timeout_seconds": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "CheckWithAccess": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Check"
          },
          {
            "$ref": "#/components/schemas/CheckAccess"
          }
        ]
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "Method": {
        "type": "string",
        "enum": [
          "GET",
          "POST",
          "PUT",
          "DELETE",
          "HEAD"
        ]
      },
      "MetricsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MetricsSummary"
          },
          {
            "type": "object",
            "required": [
              "by_region"
            ],
            "properties": {
              "by_region": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RegionMetrics"
                }
              }
            }
          }
        ]
      },
      "MetricsSummary": {
        "type": "object",
        "required": [
          "uptime_percent",
          "avg_response_time_ms",
          "p95_response_time_ms",
          "p99_response_time_ms"
        ],
        "properties": {
          "avg_response_time_ms": {
            "type": "number",
            "format": "double"
          },
          "p95_response_time_ms": {
            "type": "number",
            "format": "double"
          },
          "p99_response_time_ms": {
            "type": "number",
            "format": "double"
          },
          "uptime_percent": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PublicUser": {
        "type": "object",
        "required": [
          "user_id",
          "username"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "Region": {
        "type": "string",
        "enum": [
          "Fsn1",
          "Hel1",
          "Nbg1"
        ]
      },
      "RegionMetrics": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MetricsSummary"
          },
          {
            "type": "object",
            "required": [
              "region"
            ],
            "properties": {
              "region": {
                "$ref": "#/components/schemas/Region"
              }
            }
          }
        ]
      },
      "Vec": {
        "type": "array",
        "items": {
          "oneOf": [
            {
              "type": "object",
              "required": [
                "ServiceCheckMutation"
              ],
              "properties": {
                "ServiceCheckMutation": {
                  "type": "object",
                  "required": [
                    "check_id"
                  ],
                  "properties": {
                    "check_id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            {
              "type": "object",
              "required": [
                "ShuttingDown"
              ],
              "properties": {
                "ShuttingDown": {
                  "type": "object",
                  "required": [
                    "process_id"
                  ],
                  "properties": {
                    "process_id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          ]
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      },
      "internal_bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Health-related endpoints."
    },
    {
      "name": "users",
      "description": "User-related endpoints."
    },
    {
      "name": "checks",
      "description": "Health check management endpoints."
    },
    {
      "name": "internal",
      "description": "Internal endpoints for backend-to-backend communication."
    }
  ]
}